// apps/dashboard-tsr/src/lib/filters/types.ts
export type FilterOperator =
  | 'eq' // equals
  | 'ne' // not equals
  | 'contains' // case-insensitive substring match
  | 'notContains' // case-insensitive substring exclusion
  | 'in' // value is one of a list
  | 'notIn' // value is not in a list
  | 'gt' // greater than
  | 'gte' // greater than or equal
  | 'lt' // less than
  | 'lte' // less than or equal
  | 'between' // within an inclusive range (two values)
  | 'withinHours' // event happened within the last N hours